♻️ refactor: 기존 context 파일 코드의 관심사 분리 (#53)
